pyproject.toml
[project]
name = "mistral-ocr"
version = "0.1.0"
requires-python = ">=3.11.1, <3.12"
dependencies = [
    "click>=8.1.8",
    "jinja2>=3.1.6",
    "markdown2>=2.5.3",
    "mistralai>=1.5.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

line-length = 79
indent-size = 4

target-version = "py311"
